requirements.txt
uvicorn
uvicorn[standard]
fastapi
python-dotenv==1.0.0
openai==1.13.3